sh script

+++ 0x06-signals/8-signal_send.sh
@@ -1,3 +1,9 @@
-#!/usr/bin/env bash
-(( $# != 1 )) && { echo "Usage: $0 <pid>" >&2; exit 1; }
+#!/usr/bin/env sh
+if [ $# != 1 ]; then 
+{ 
+    echo "Usage: $0 <pid>" >&2
+    exit 1
+}
+fi
+
 kill -3 "$1"
